Wrote a reader
I wrote a reader to get past quarterly returns from a csv file.

In my next commit, I plan to write a portfolio class.

I also want to write a "Simulate quarter" method which will accept a
portfolio object and a Returns vector. It will randomly select one
element of the returns vector and use it to update the portfolio.