Added settings file + Parser
I added some file/string operations settings file parser. These are
literally straight from my SPH code (but with a dedicated comment line
detector, and some other minor tweaks). This is in IO_Ops.cxx/hxx

I also wrote a settings parser for the settings file.

I also renamed Reader as Data_Reader and moved it to the newly created
IO directory (for all IO stuff).

Everything seems to work now :)

I forgot to add a setting for the number of quarters we should simulate
for, as well as the number of simulations (I need to add simulation
settings... oops), so expect that in the next commit.

I also need to write code to run lots of simulations and then produce
stats on the results of those simulations. Expect that soonish.

Otherwise, this project is mostly done (in terms of the code bits).